create style folder and moved all css files in it
