+++ src/styles/Trips.css
@@ -1,0 +1,3 @@
+h3 {
+  letter-spacing: 4px;
+}
